+++ tsrc/globals/gameSocket.ts
@@ -6,6 +6,7 @@ import fs from "fs";
 import { gameState, GameState, MenuStates } from "./gameState";
 import { settings } from "./settings";
 import { isInt } from "../utility";
+import { lobbyControl } from "./../modules/lobbyControl";
 
 import { warControl } from "./warControl";
 import { GameClientLobbyPayload, Regions } from "wc3mt-lobby-container";
@@ -111,6 +112,7 @@ class GameSocket extends Global {
   gameWebSocket: WebSocket | null = null;
   voteTimer: NodeJS.Timeout | null = null;
   gameState = gameState;
+  lobbyControl = lobbyControl;
   sendingInGameChat: { active: boolean; queue: Array<string> } = {
     active: false,
     queue: [],
@@ -135,7 +137,11 @@ class GameSocket extends Global {
     });
     this.gameWebSocket.on("message", (data) => {
       let parsedData = JSON.parse(data.toString());
-      this.emitEvent(parsedData);
+      if (parsedData.messageType === "ChatMessage") {
+        this.handleChatMessage(parsedData.payload);
+      } else {
+        this.emitEvent(parsedData);
+      }
       switch (parsedData.messageType) {
         case "ScreenTransitionInfo":
           this.gameState.updateGameState({ screenState: parsedData.payload.screen });
@@ -147,14 +153,6 @@ class GameSocket extends Global {
           } else {
             this.handleGlueScreen("CUSTOM_LOBBIES");
           }
-          break;
-        case "OnChannelUpdate":
-          if (parsedData.payload.gameChat) {
-            //console.log(data.payload);
-          }
-          break;
-        case "ChatMessage":
-          this.handleChatMessage(parsedData.payload);
           break;
         case "MultiplayerGameLeave":
           clearLobby();
@@ -273,7 +271,7 @@ class GameSocket extends Global {
       ) {
         var sender = this.gameState.values.selfBattleTag;
       } else {
-        let possiblePlayers = this.lobby?.microLobby?.searchPlayer(
+        let possiblePlayers = this.lobbyControl?.microLobby?.searchPlayer(
           payload.message.sender
         );
         if (possiblePlayers && possiblePlayers.length === 1) {
@@ -303,668 +301,23 @@ class GameSocket extends Global {
           return;
         }
       }
+      // Message has a sender and is probably not a debug message.
+      if (!this.lobbyControl?.microLobby?.newChat(sender, payload.message.content)) {
+        // Filter out repeated messages sent w/in 1 second
+        // TODO: more spam filters
+        return;
+      }
+      this.emitEvent({
+        ChatMessage: {
+          message: { sender, content: payload.message.content, source: "gameChat" },
+        },
+      });
       {
-        if (!this.lobby?.microLobby?.newChat(sender, payload.message.content)) {
-          // Filter out repeated messages sent w/in 1 second
-          // TODO: more spam filters
-          return;
-        }
         if (
           sender !== this.gameState.values.selfBattleTag &&
           payload.message.content.match(/^!debug/)
         ) {
-          this.lobby?.banPlayer(sender);
-        } else if (payload.message.content.match(/^\?votestart$/i)) {
-          if (
-            settings.values.autoHost.voteStart &&
-            this.lobby?.voteStartVotes &&
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            ["rapidHost", "smartHost"].includes(settings.values.autoHost.type)
-          ) {
-            if (!this.lobby?.microLobby?.allPlayers.includes(sender)) {
-              this.sendChatMessage("Only players may vote start.");
-              return;
-            }
-            if (this.lobby?.voteStartVotes.length === 0) {
-              if (
-                (settings.values.autoHost.voteStartTeamFill &&
-                  this.lobby?.allPlayerTeamsContainPlayers()) ||
-                !settings.values.autoHost.voteStartTeamFill
-              ) {
-                this.voteTimer = setTimeout(cancelVote, 60000);
-                this.sendChatMessage("You have 60 seconds to ?votestart.");
-              } else {
-                this.sendChatMessage("Unavailable. Not all teams have players.");
-                return;
-              }
-            }
-            if (!this.lobby?.voteStartVotes.includes(sender) && this.voteTimer) {
-              this.lobby?.voteStartVotes.push(sender);
-              if (
-                this.lobby?.voteStartVotes.length >=
-                this.lobby?.microLobby?.nonSpecPlayers.length *
-                  (settings.values.autoHost.voteStartPercent / 100)
-              ) {
-                this.info("Vote start succeeded");
-                this.startGame();
-              } else {
-                this.sendChatMessage(
-                  Math.ceil(
-                    this.lobby?.microLobby?.nonSpecPlayers.length *
-                      (settings.values.autoHost.voteStartPercent / 100) -
-                      this.lobby?.voteStartVotes.length
-                  ).toString() + " more vote(s) required."
-                );
-              }
-            }
-          }
-        } else if (payload.message.content.match(/^\?stats/)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic?.isHost &&
-            settings.values.elo.type !== "off" &&
-            this.lobby?.microLobby?.statsAvailable
-          ) {
-            let data: false | PlayerData;
-            let playerTarget = payload.message.content.split(" ")[1];
-            if (playerTarget) {
-              let targets = this.lobby?.microLobby?.searchPlayer(playerTarget);
-              if (targets.length === 1) {
-                sender = targets[0];
-                data = this.lobby?.getPlayerData(sender);
-              } else if (targets.length > 1) {
-                this.sendChatMessage("Multiple players found. Please be more specific.");
-                return;
-              } else {
-                this.sendChatMessage("No player found.");
-                return;
-              }
-            } else {
-              data = this.lobby?.getPlayerData(sender);
-            }
-            if (data) {
-              if (!data.extra || data.extra?.rating === -1) {
-                this.sendChatMessage("Data pending");
-              } else {
-                this.sendChatMessage(
-                  sender +
-                    " ELO: " +
-                    data.extra.rating +
-                    ", Rank: " +
-                    data.extra.rank +
-                    ", Played: " +
-                    data.extra.played +
-                    ", Wins: " +
-                    data.extra.wins +
-                    ", Losses: " +
-                    data.extra.losses +
-                    ", Last Change: " +
-                    data.extra.lastChange
-                );
-              }
-            } else {
-              this.sendChatMessage("No data available or pending?");
-            }
-          } else {
-            this.sendChatMessage("Data not available");
-          }
-        } else if (payload.message.content.match(/^\?sp$/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            this.lobby?.shufflePlayers();
-          }
-        } else if (payload.message.content.match(/^\?st$/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic?.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            this.lobby?.shufflePlayers(false);
-          }
-        } else if (payload.message.content.match(/^\?start$/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            this.startGame();
-          }
-        } else if (payload.message.content.match(/^\?a$/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            this.cancelStart();
-          }
-        } else if (payload.message.content.match(/^\?closeall$/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            this.sendChatMessage("!closeall");
-          }
-        } else if (payload.message.content.match(/^\?hold$/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            let targetPlayer = payload.message.content.split(" ")[1];
-            if (targetPlayer) {
-              this.sendChatMessage("!hold " + targetPlayer);
-            } else {
-              this.sendChatMessage("Player target required.");
-            }
-          }
-        } else if (payload.message.content.match(/^\?mute$/i)) {
-          if (banWhiteListSingle.checkRole(sender, "moderator")) {
-            let targetPlayer = payload.message.content.split(" ")[1];
-            if (targetPlayer) {
-              this.sendChatMessage("!mute " + targetPlayer);
-              this.info(sender + " muted " + targetPlayer);
-            } else {
-              this.sendChatMessage("Player target required.");
-            }
-          }
-        } else if (payload.message.content.match(/^\?unmute$/i)) {
-          if (banWhiteListSingle.checkRole(sender, "moderator")) {
-            let targetPlayer = payload.message.content.split(" ")[1];
-            if (targetPlayer) {
-              this.sendChatMessage("!unmute " + targetPlayer);
-              this.info(sender + " unmuted " + targetPlayer);
-            } else {
-              this.sendChatMessage("Player target required.");
-            }
-          }
-        } else if (payload.message.content.match(/^\?openall$/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            this.sendChatMessage("!openall");
-          }
-        } else if (payload.message.content.match(/^\?swap/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "baswapper")
-          ) {
-            let [command, ...args] = payload.message.content.split(" ");
-            if (args.length === 2) {
-              let playerData = this.lobby?.microLobby?.getAllPlayerData();
-              let tenMinutesAgo = Date.now() - 10 * 60 * 1000;
-              if (isInt(args[1], 24, 1) && isInt(args[0], 24, 1)) {
-                if (
-                  banWhiteListSingle.checkRole(sender, "swapper") ||
-                  (playerData[this.lobby?.microLobby?.slots[parseInt(args[0]) - 1].name]
-                    .joinedAt > tenMinutesAgo &&
-                    playerData[this.lobby?.microLobby?.slots[parseInt(args[1]) - 1].name]
-                      .joinedAt > tenMinutesAgo)
-                ) {
-                  this.lobby?.swapPlayers({
-                    slots: [
-                      ensureInt(args[0]) as SlotNumbers,
-                      ensureInt(args[1]) as SlotNumbers,
-                    ],
-                  });
-                } else {
-                  this.sendChatMessage(
-                    "You can only swap players who joined within the last 10 minutes."
-                  );
-                }
-              } else if (
-                this.lobby?.microLobby?.searchPlayer(args[1]).length === 1 &&
-                this.lobby?.microLobby?.searchPlayer(args[0]).length === 1
-              ) {
-                if (
-                  banWhiteListSingle.checkRole(sender, "swapper") ||
-                  (playerData[this.lobby?.microLobby?.searchPlayer(args[1])[0]].joinedAt >
-                    tenMinutesAgo &&
-                    playerData[this.lobby?.microLobby?.searchPlayer(args[0])[0]]
-                      .joinedAt > tenMinutesAgo)
-                ) {
-                  this.lobby?.swapPlayers({ players: [args[0], args[1]] });
-                } else {
-                  this.sendChatMessage(
-                    "You can only swap players who joined within the last 10 minutes."
-                  );
-                }
-              } else {
-                this.sendChatMessage("All swap players not found, or too many matches.");
-              }
-            } else {
-              this.sendChatMessage("Invalid swap arguments");
-            }
-          }
-        } else if (payload.message.content.match(/^\?handi/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            if (payload.message.content.split(" ").length === 3) {
-              var target = payload.message.content.split(" ")[1];
-              var handicap = parseInt(payload.message.content.split(" ")[2]);
-              if (handicap) {
-                if (isInt(target, 24, 1)) {
-                  this.lobby?.setHandicapSlot(parseInt(target) - 1, handicap);
-                } else {
-                  this.lobby?.setPlayerHandicap(target, handicap);
-                }
-              } else {
-                this.sendChatMessage("Invalid handicap");
-              }
-            } else {
-              this.sendChatMessage("Invalid number of arguments");
-            }
-          }
-        } else if (payload.message.content.match(/^\?close/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            var target = payload.message.content.split(" ")[1];
-            if (target) {
-              if (isInt(target, 24, 1)) {
-                this.lobby?.closeSlot(parseInt(target) - 1);
-              } else {
-                let targets = this.lobby?.microLobby?.searchPlayer(target);
-                if (targets.length === 1) {
-                  this.lobby?.closePlayer(targets[0]);
-                } else if (targets.length > 1) {
-                  this.sendChatMessage(
-                    "Multiple matches found. Please be more specific."
-                  );
-                } else {
-                  this.sendChatMessage("No matches found.");
-                }
-              }
-            } else {
-              this.sendChatMessage("Kick target required");
-            }
-          }
-        } else if (payload.message.content.match(/^\?open/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            var target = payload.message.content.split(" ")[1];
-            if (target) {
-              if (isInt(target, 24, 1)) {
-                this.lobby?.openSlot(parseInt(target) - 1);
-              } else {
-                let targets = this.lobby?.microLobby?.searchPlayer(target);
-                if (targets.length === 1) {
-                  this.lobby?.kickPlayer(targets[0]);
-                } else if (targets.length > 1) {
-                  this.sendChatMessage(
-                    "Multiple matches found. Please be more specific."
-                  );
-                } else {
-                  this.sendChatMessage("No matches found.");
-                }
-              }
-            } else {
-              this.sendChatMessage("Kick target required");
-            }
-          }
-        } else if (payload.message.content.match(/^\?kick/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            var target = payload.message.content.split(" ")[1];
-            if (target) {
-              if (isInt(target, 24, 1)) {
-                this.lobby?.kickSlot(parseInt(target) - 1);
-              } else {
-                let targets = this.lobby?.microLobby?.searchPlayer(target);
-                if (targets.length === 1) {
-                  this.lobby?.kickPlayer(targets[0]);
-                } else if (targets.length > 1) {
-                  this.sendChatMessage(
-                    "Multiple matches found. Please be more specific."
-                  );
-                } else {
-                  this.sendChatMessage("No matches found.");
-                }
-              }
-            } else {
-              this.sendChatMessage("Kick target required");
-            }
-          }
-        } else if (payload.message.content.match(/^\?ban/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            var targetPlayer = payload.message.content.split(" ")[1];
-            if (targetPlayer) {
-              var reason = payload.message.content.split(" ").slice(2).join(" ") || "";
-              if (isInt(targetPlayer, 24, 1)) {
-                this.lobby?.banSlot(parseInt(targetPlayer) - 1);
-                banWhiteListSingle.banPlayer(
-                  this.lobby?.microLobby?.slots[targetPlayer].name,
-                  sender,
-                  this.lobby?.microLobby?.region,
-                  reason
-                );
-              } else {
-                if (targetPlayer.match(/^\D\S{2,11}#\d{4,8}$/)) {
-                  this.sendChatMessage("Banning out of lobby player.");
-                  banWhiteListSingle.banPlayer(
-                    targetPlayer,
-                    sender,
-                    this.lobby?.microLobby?.region,
-                    reason
-                  );
-                } else {
-                  let targets = this.lobby?.microLobby?.searchPlayer(targetPlayer);
-                  if (targets.length === 1) {
-                    banWhiteListSingle.banPlayer(
-                      targets[0],
-                      sender,
-                      this.lobby?.microLobby?.region,
-                      reason
-                    );
-                  } else if (targets.length > 1) {
-                    this.sendChatMessage(
-                      "Multiple matches found. Please be more specific."
-                    );
-                  } else {
-                    this.sendChatMessage("No matches found.");
-                  }
-                }
-              }
-            } else {
-              this.sendChatMessage("Target required");
-            }
-          }
-        } else if (payload.message.content.match(/^\?unban/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            var target = payload.message.content.split(" ")[1];
-            if (target) {
-              if (target.match(/^\D\S{2,11}#\d{4,8}$/)) {
-                this.sendChatMessage("Unbanning out of lobby player.");
-                banWhiteListSingle.unBanPlayer(target, sender);
-              } else {
-                this.sendChatMessage("Full battleTag required");
-                this.info("Full battleTag required");
-              }
-            } else {
-              this.sendChatMessage("Ban target required");
-              this.info("Ban target required");
-            }
-          }
-        } else if (payload.message.content.match(/^\?white/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic?.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            var targetPlayer = payload.message.content.split(" ")[1];
-            if (targetPlayer) {
-              var reason = payload.message.content.split(" ").slice(2).join(" ") || "";
-              if (isInt(targetPlayer, 24, 1)) {
-                banWhiteListSingle.whitePlayer(
-                  this.lobby?.microLobby?.slots[targetPlayer].name,
-                  sender,
-                  this.lobby?.microLobby?.region,
-                  reason
-                );
-              } else {
-                if (targetPlayer.match(/^\D\S{2,11}#\d{4,8}$/)) {
-                  this.sendChatMessage("Whitelisting out of lobby player.");
-                  banWhiteListSingle.whitePlayer(
-                    targetPlayer,
-                    sender,
-                    this.lobby?.microLobby?.region,
-                    reason
-                  );
-                } else {
-                  let targets = this.lobby?.microLobby?.searchPlayer(targetPlayer);
-                  if (targets.length === 1) {
-                    banWhiteListSingle.whitePlayer(
-                      targets[0],
-                      sender,
-                      this.lobby?.microLobby?.region,
-                      reason
-                    );
-                  } else if (targets.length > 1) {
-                    this.sendChatMessage(
-                      "Multiple matches found. Please be more specific."
-                    );
-                  } else {
-                    this.sendChatMessage("No matches found.");
-                  }
-                }
-              }
-            } else {
-              this.sendChatMessage("Target required");
-            }
-          }
-        } else if (payload.message.content.match(/^\?unwhite/i)) {
-          // TODO: In lobby search and removal
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "moderator")
-          ) {
-            var target = payload.message.content.split(" ")[1];
-            if (target) {
-              if (target.match(/^\D\S{2,11}#\d{4,8}$/)) {
-                this.sendChatMessage("Un-whitelisting out of lobby player.");
-                banWhiteListSingle.unWhitePlayer(target, sender);
-              } else {
-                this.sendChatMessage("Full battleTag required");
-                this.info("Full battleTag required");
-              }
-            } else {
-              this.sendChatMessage("Un-whitelist target required");
-              this.info("Un-whitelist target required");
-            }
-          }
-        } else if (payload.message.content.match(/^\?perm/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "admin")
-          ) {
-            var target = payload.message.content.split(" ")[1];
-            var perm: "mod" | "baswapper" | "swapper" | "moderator" | "admin" =
-              (payload.message.content.split(" ")[2]?.toLowerCase() as
-                | null
-                | "baswapper"
-                | "swapper"
-                | "moderator"
-                | "admin") ?? "mod";
-            perm = perm === "mod" ? "moderator" : perm;
-            if (target) {
-              if (["baswapper", "swapper", "moderator", "admin"].includes(perm)) {
-                if (target.match(/^\D\S{2,11}#\d{4,8}$/)) {
-                  this.sendChatMessage("Assigning out of lobby player " + perm + ".");
-                  banWhiteListSingle.addAdmin(
-                    target,
-                    sender,
-                    this.lobby?.microLobby?.region,
-                    perm
-                  );
-                } else {
-                  let targets = this.lobby?.microLobby?.searchPlayer(target);
-                  if (targets.length === 1) {
-                    if (
-                      banWhiteListSingle.addAdmin(
-                        targets[0],
-                        sender,
-                        this.lobby?.microLobby?.region,
-                        perm
-                      )
-                    ) {
-                      this.sendChatMessage(
-                        targets[0] + " has been promoted to " + perm + "."
-                      );
-                    } else {
-                      this.sendChatMessage(
-                        "Could not promote " + targets[0] + " to " + perm + "."
-                      );
-                    }
-                  } else if (targets.length > 1) {
-                    this.sendChatMessage(
-                      "Multiple matches found. Please be more specific."
-                    );
-                  } else {
-                    this.sendChatMessage("No matches found.");
-                  }
-                }
-              } else {
-                this.sendChatMessage("Invalid permission");
-              }
-            } else {
-              this.sendChatMessage("Target required");
-            }
-          }
-        } else if (payload.message.content.match(/^\?unperm/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic?.isHost &&
-            banWhiteListSingle.checkRole(sender, "admin")
-          ) {
-            var target = payload.message.content.split(" ")[1];
-            if (target) {
-              if (target.match(/^\D\S{2,11}#\d{4,8}$/)) {
-                if (banWhiteListSingle.removeAdmin(target, sender)) {
-                  this.sendChatMessage(
-                    "Removed perm from out of lobby player: " + target
-                  );
-                } else {
-                  this.sendChatMessage(
-                    "Could not remove perm from out of lobby player: " + target
-                  );
-                }
-              } else {
-                let targets = this.lobby?.microLobby?.searchPlayer(target);
-                if (targets.length === 1) {
-                  if (banWhiteListSingle.removeAdmin(targets[0], sender)) {
-                    this.sendChatMessage(targets[0] + " has been demoted.");
-                  } else {
-                    this.sendChatMessage(targets[0] + " has no permissions.");
-                  }
-                } else if (targets.length > 1) {
-                  this.sendChatMessage(
-                    "Multiple matches found. Please be more specific."
-                  );
-                } else {
-                  this.sendChatMessage("No matches found.");
-                }
-              }
-            } else {
-              this.sendChatMessage("Target required");
-            }
-          }
-        } else if (payload.message.content.match(/^\?autohost/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "admin")
-          ) {
-            var target = payload.message.content.split(" ")[1];
-            if (target) {
-              target = target.toLowerCase();
-              if (["off", "rapid", "lobby", "smart"].includes(target)) {
-                if (target !== "off") {
-                  target += "Host";
-                }
-                this.sendChatMessage("Setting autohost type to: " + target);
-                settings.updateSettings({
-                  autoHost: { type: target as AutoHostSettings["type"] },
-                });
-              } else {
-                this.sendChatMessage("Invalid autohost type");
-              }
-            } else {
-              this.sendChatMessage(
-                "Autohost current type: " + settings.values.autoHost.type
-              );
-            }
-          } else {
-            this.sendChatMessage("You do not have permission to use this command.");
-          }
-        } else if (payload.message.content.match(/^\?autostart/i)) {
-          if (
-            this.lobby?.microLobby?.lobbyStatic.isHost &&
-            banWhiteListSingle.checkRole(sender, "admin")
-          ) {
-            var target = payload.message.content.split(" ")[1];
-            if (target) {
-              if (isInt(target, 24, 0)) {
-                var startTarget = parseInt(target);
-                this.sendChatMessage(
-                  "Setting autostart number to: " + startTarget.toString()
-                );
-                if (settings.values.autoHost.type === "off") {
-                  this.sendChatMessage("Autohost must be enabled to autostart.");
-                }
-                settings.updateSettings({ autoHost: { minPlayers: startTarget } });
-              } else {
-                this.sendChatMessage("Invalid autostart number");
-              }
-            } else {
-              this.sendChatMessage(
-                "Autostart current number: " + settings.values.autoHost.minPlayers
-              );
-            }
-          } else {
-            this.sendChatMessage("You do not have permission to use this command.");
-          }
-        } else if (payload.message.content.match(/^\?(help)|(commands)/i)) {
-          if (this.lobby?.microLobby?.lobbyStatic.isHost) {
-            if (this.lobby?.microLobby?.statsAvailable) {
-              this.sendChatMessage(
-                "?stats <?player>: Return back your stats, or target player stats"
-              );
-            }
-            if (
-              ["rapidHost", "smartHost"].includes(settings.values.autoHost.type) &&
-              settings.values.autoHost.voteStart
-            ) {
-              this.sendChatMessage("?voteStart: Starts or accepts a vote to start");
-            }
-            if (banWhiteListSingle.checkRole(sender, "moderator")) {
-              this.sendChatMessage("?a: Aborts game start");
-              this.sendChatMessage(
-                "?ban <name|slotNumber> <?reason>: Bans a player forever"
-              );
-              this.sendChatMessage(
-                "?close<?all> <name|slotNumber>: Closes all / a slot/player"
-              );
-              this.sendChatMessage(
-                "?handi <name|slotNumber> <50|60|70|80|100>: Sets slot/player handicap"
-              );
-              this.sendChatMessage("?hold <name>: Holds a slot");
-              this.sendChatMessage(
-                "?kick <name|slotNumber> <?reason>: Kicks a slot/player"
-              );
-              this.sendChatMessage("?<un>mute <player>: Mutes/un-mutes a player");
-              this.sendChatMessage(
-                "?open<?all> <name|slotNumber> <?reason>: Opens all / a slot/player"
-              );
-              this.sendChatMessage("?unban <name>: Un-bans a player");
-              this.sendChatMessage("?white <name>: Whitelists a player");
-              this.sendChatMessage("?unwhite <name>: Un-whitelists a player");
-              this.sendChatMessage("?start: Starts game");
-              this.sendChatMessage(
-                "?swap <name|slotNumber> <name|slotNumber>: Swaps players"
-              );
-              this.sendChatMessage("?sp: Shuffles players completely randomly");
-              this.sendChatMessage("?st: Shuffles players randomly between teams");
-            }
-            if (banWhiteListSingle.checkRole(sender, "admin")) {
-              this.sendChatMessage(
-                "?perm <name> <?admin|mod|swapper>: Promotes a player to a role (mod by default)"
-              );
-              this.sendChatMessage("?unperm <name>: Demotes player to normal");
-              this.sendChatMessage(
-                "?autohost <?off|rapid|lobby|smart>: Gets/?Sets autohost type"
-              );
-            }
-            this.sendChatMessage(
-              "?help: Shows commands with <required arg> <?optional arg>"
-            );
-          }
+          this.lobbyControl?.banPlayer(sender);
         }
         var translatedMessage = "";
         if (payload.message.content.length > 4) {
